[issue439] use int for vars and values

+++ src/search/relaxation_heuristic.cc
@@ -31,11 +31,11 @@ void RelaxationHeuristic::initialize() {
     // Build propositions.
     int prop_id = 0;
     Variables vars = task.get_variables();
-    size_t num_vars = vars.size();
+    int num_vars = vars.size();
     propositions.resize(num_vars);
-    for (size_t var = 0; var < num_vars; var++) {
-        size_t num_values = vars[var].get_domain_size();
-        for (size_t value = 0; value < num_values; value++)
+    for (int var = 0; var < num_vars; var++) {
+        int num_values = vars[var].get_domain_size();
+        for (int value = 0; value < num_values; value++)
             propositions[var].push_back(Proposition(prop_id++));
     }
